+++ lib/public/Model.js
@@ -30,7 +30,7 @@ import { AboutModel } from './views/About/About.js';
 import { StatisticsPageModel } from './views/Statistics/StatisticsPageModel.js';
 import { LhcPeriodsModel } from './views/lhcPeriods/LhcPeriodsModel.js';
 import { HomePageModel } from './views/Home/Overview/HomePageModel.js';
-import { DataPassesModel } from './views/DataPasses/DataPassesPerLhcPeriodsModel.js';
+import { DataPassesModel } from './views/DataPasses/DataPassesModel.js';
 
 /**
  * Root of model tree
